requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.3
packaging==23.2
PyPDF2==3.0.1
python-docx==1.1.0
typing_extensions==4.9.0
Werkzeug==3.0.1
requests==2.28.1

